// backendlib/workflow/types.ts
export const wfInstSchemaId = '/workflow/WfInstance';
// ...
export interface MappedTrigger extends BaseTrigger {
  processingSteps?: MappedProcessingStep[];
  transitions: MappedTransition[];
}
// ...
export interface UserSession {
  userId: string;
  roles?: string[];
  org?: string;
  name?: string;
}
// ...
export interface MappedStateType {
  name: string;
  isStartState: boolean;
  isEndState: boolean;
  triggers: MappedTrigger[];
  description?: string;
}

export interface MappedProcessingStep {
  name: string;
  description?: string;
  importPath: string;
  func: (ctx: WorkflowContext) => Promise<WorkflowContext | void> | WorkflowContext | void;
}

export interface MappedWorkflowDefinition {
  name: string;
  startStateName: string;
  description?: string;
  domainSchemaId: string;
  states: MappedStateType[];
}

export interface WorkflowStartHandlerInput {
  org?: string;
  entityIds?: any;
  instanceId?: string; // new workflow ID
  wfName: string; // name of the workflow
  domainObjectId?: string;
  domainSchemaId?: string;
  eventName: string; // Name of the event specified in start state trigger
  eventData: any; // payload "formData"
}